feat: add defineConfig helper and /config submodule
Expose a Vite-style `defineConfig` from `@ilbertt/bun-sqlgen/config` so
`sqlgen.config.ts` is type-checked and its callbacks inferred without a manual
annotation. The submodule is self-contained (imports only @electric-sql/pglite),
so it builds with the lib and ships a declaration free of the private core
package; core now derives its config-loader type from IntrospectorOptions.
