+++ examples/with-config/sqlgen.config.ts
@@ -1,9 +1,12 @@
 import { citext } from '@electric-sql/pglite/contrib/citext';
+import { defineConfig } from '@repo/bun-sqlgen/config';
 
 // Shapes the throwaway PGlite introspection DB so the migrations apply exactly as
 // they would against production. Auto-discovered by `bun-sqlgen` because it sits in
-// the directory codegen runs from (or point at it with `--config`).
-export default {
+// the directory codegen runs from (or point at it with `--config`). `defineConfig`
+// is a no-op at runtime — it just type-checks the object and infers the callback
+// parameters below.
+export default defineConfig({
   // Make the `citext` extension available, so a migration's `CREATE EXTENSION
   // citext` can succeed. Real Postgres has it on disk; PGlite has to be handed it.
   extensions: () => ({ citext }),
@@ -19,6 +22,5 @@ export default {
   // Rewrite each migration before it is applied. A multi-statement file runs in a
   // transaction, and `CREATE INDEX CONCURRENTLY` cannot — production needs it to
   // avoid locking the table, the throwaway DB does not, so strip it.
-  transformMigration: ({ sql }: { sql: string; filename: string }) =>
-    sql.replace(/\bCONCURRENTLY\b/g, ''),
-};
+  transformMigration: ({ sql }) => sql.replace(/\bCONCURRENTLY\b/g, ''),
+});
