+++ packages/bun-sqlgen-core/src/index.ts
@@ -14,12 +14,9 @@ export type {
   DiscoveredQuery,
   NullabilityReason,
   Overrides,
-  PostgresConfig,
   Provenance,
   ResolvedField,
   ResultField,
-  SqlgenConfig,
-  SqliteConfig,
   TypeCatalog,
   TypeInfo,
   WritableColumns,
