+++ packages/bun-sqlgen/package.json
@@ -9,6 +9,10 @@
     ".": {
       "types": "./src/index.ts",
       "bun": "./src/index.ts"
+    },
+    "./config": {
+      "types": "./src/config.ts",
+      "bun": "./src/config.ts"
     }
   },
   "scripts": {
